Esure accessors are Float64 (true for all except new bladeRF)

+++ src/Assessors.jl
@@ -21,22 +21,22 @@ getTimestamp(obj::BladeRFBinding) = BladeRFBindings.getTimestamp(obj);
 Get the current sampling rate of the radio device 
 The second parameter (optionnal) speicfies the Rx or Tx board (default : Rx)
 """ 
-getSamplingRate(obj::AbstractSDR;mode=:rx) = ((mode == :rx) ? obj.rx.samplingRate : obj.tx.samplingRate)
-getSamplingRate(obj::PlutoSDR;mode=:rx) = ((mode == :rx) ? obj.rx.effectiveSamplingRate : obj.tx.effectiveSamplingRate)
+getSamplingRate(obj::AbstractSDR;mode=:rx) = ((mode == :rx) ? Float64(obj.rx.samplingRate) : Float64(obj.tx.samplingRate))
+getSamplingRate(obj::PlutoSDR;mode=:rx) = ((mode == :rx) ? Float64(obj.rx.effectiveSamplingRate) : Float64(obj.tx.effectiveSamplingRate))
 
 """ 
 Get the current carrier frequency   of the radio device 
 The second parameter (optionnal) speicfies the Rx or Tx board (default : Rx)
 """ 
-getCarrierFreq(obj::AbstractSDR;mode=:rx) = (mode == :rx) ? obj.rx.carrierFreq : obj.tx.carrierFreq
-getCarrierFreq(obj::PlutoSDR;mode=:rx) = (mode == :rx) ? obj.rx.effectiveCarrierFreq : obj.tx.effectiveCarrierFreq
+getCarrierFreq(obj::AbstractSDR;mode=:rx) = (mode == :rx) ? Float64(obj.rx.carrierFreq) : Float64(obj.tx.carrierFreq)
+getCarrierFreq(obj::PlutoSDR;mode=:rx) = (mode == :rx) ? Float64(obj.rx.effectiveCarrierFreq) : Float64(obj.tx.effectiveCarrierFreq)
 
 
 """ 
 Get the current radio gain 
 The second parameter (optionnal) specifies the Rx or Tx board (default : Rx)
 """
-getGain(obj::AbstractSDR;mode=:rx) = (mode == :rx) ? obj.rx.gain : obj.tx.gain 
+getGain(obj::AbstractSDR;mode=:rx) = (mode == :rx) ? Float64(obj.rx.gain) : Float64(obj.tx.gain) 
 getGain(obj::PlutoSDR;mode=:rx) = AdalmPluto.getGain(obj)
